Append "redis" group rather than making it exclusive
In our environment, we have a redis sidecar container in each of our pods,
using a unix domain socket shared amongst multiple containers.

To deal with the arbitrary UID problem, we typically give +rw access to the
"root" group. However, we ran into issues with the Redis container, since the
user only belongs to the "redis" group. This PR fixes that problem for us.